fix(frontend): extract profile sign-in label helpers out of page.tsx (#855)
Next.js 16 App Router rejects arbitrary named exports from a page file
("not a valid Page export field"). profile/page.tsx exported
getSignInMethodLabel and getRefreshProviderName, which failed the build-time
type check — previously masked because the Turbopack abort happened first.

Move both pure i18n helpers to a sibling signInLabels.ts; page.tsx and
page.test.tsx import from there. No behavior change (23 profile tests pass).
